+++ test/specs/assets/misc/validator-collection.spec.js
@@ -1,0 +1,1 @@
+//NOTE: tested through the form stuff